requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Browser Automation
playwright>=1.40.0

# LLM Integration
openai>=1.3.7

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
pypdf>=4.0.0

# Web Scraping
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.1
pillow>=10.1.0
pydantic[email]